+++ ChangeLog.md
@@ -6,3 +6,4 @@ v 0.1.1.0
 * Added export of newPandoc and NamedDoc to Knit.Report to facilitate multi-doc use.
 * Removed redundant imports from Knit.Report.Input.Table.Colonnade
 * Added a multi-doc example
+* (internal) Removed odd, and no longer necessary, "LastMember" constraint from knit functions.